+++ src/reducers/index.js
@@ -5,12 +5,12 @@ import { reducer as formReducer } from 'redux-form';
 
 import reports from 'sections/Reports/reducers';
 import settings from 'sections/Settings/reducers';
+import projects from 'sections/Projects/reducers';
 
 import runMigrations from '../migrations';
 
 import account from './account';
 import app from './app';
-import projects from './projects';
 import time from './time';
 
 const combinedReducers = combineReducers({
